ambulance_backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==10.4.0
